Move ResolvedRequirement to `pex.resolve` package.
Also re-shuffle related types and clean up the Locker interaction.
